# admin_blog/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "思琼碧落的博客系统 API",
    "description": "个人博客系统的完整REST API文档，支持用户管理、文章发布、分类管理、音乐分享和系统配置等功能。v2.1.0更新：修复了用户角色和配置类型定义，补充了缺失的API接口。",
    "version": "2.1.0",
    "contact": {
      "name": "思琼碧落",
      "email": "zzb@siqiongbiluo.love",
      "url": "https://siqiongbiluo.love"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:9000/api",
      "description": "本地开发环境"
    },
    {
      "url": "https://siqiongbiluo.love/api",
      "description": "生产环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT认证，请在请求头中添加：Authorization: Bearer {token}"
      }
    },
    "schemas": {
...
      "File": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "文件ID"
          },
          "file_name": {
            "type": "string",
            "description": "文件名",
            "example": "1734567890_document.pdf"
          },
          "original_name": {
            "type": "string",
            "description": "原始文件名",
            "example": "重要文档.pdf"
          },
          "file_path": {
            "type": "string",
            "description": "文件路径",
            "example": "./files/1734567890_document.pdf"
          },
          "file_url": {
            "type": "string",
            "description": "访问URL",
            "example": "/static/files/1734567890_document.pdf"
          },
          "file_size": {
            "type": "integer",
            "description": "文件大小（字节）",
            "example": 2048000
          },
          "mime_type": {
            "type": "string",
            "description": "MIME类型",
            "example": "application/pdf"
          },
          "file_type": {
            "type": "string",
            "description": "文件类型",
            "enum": ["document", "video", "audio", "archive", "other"],
            "example": "document"
          },
          "description": {
            "type": "string",
            "description": "文件描述",
            "nullable": true,
            "example": "这是一个重要的技术文档"
          },
          "tags": {
            "type": "array",
            "description": "文件标签",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "example": ["技术", "文档", "重要"]
          },
          "uploader_id": {
            "type": "integer",
            "description": "上传者ID"
          },
          "uploader_username": {
            "type": "string",
            "description": "上传者用户名（关联查询）",
            "example": "admin"
          },
          "related_id": {
            "type": "integer",
            "description": "关联ID（文章ID、分类ID等）",
            "nullable": true,
            "example": 123
          },
          "download_count": {
            "type": "integer",
            "description": "下载次数",
            "default": 0,
            "example": 50
          },
          "status": {
            "type": "integer",
            "description": "状态：0-删除，1-正常",
            "enum": [0, 1],
            "default": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "操作成功"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "操作失败"
          },
          "error": {
            "type": "string",
            "description": "错误详情"
          }
        }
      },
      "PaginationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "pagination": {
                "type": "object",
                "properties": {
                  "current": {
                    "type": "integer",
                    "description": "当前页码",
                    "example": 1
                  },
                  "pageSize": {
                    "type": "integer",
                    "description": "每页数量",
                    "example": 10
                  },
                  "total": {
                    "type": "integer",
                    "description": "总记录数",
                    "example": 100
                  },
                  "pages": {
                    "type": "integer",
                    "description": "总页数",
                    "example": 10
                  }
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/files": {
      "get": {
        "tags": ["文件管理"],
        "summary": "获取文件列表",
        "description": "获取文件列表，支持分页和筛选（需要管理员权限）",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {"type": "integer", "minimum": 1, "default": 1}
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页数量",
            "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 20}
          },
          {
            "name": "file_type",
            "in": "query",
            "description": "文件类型筛选",
            "schema": {"type": "string", "enum": ["document", "video", "audio", "archive", "other"]}
          },
          {
            "name": "uploader_id",
            "in": "query",
            "description": "上传者ID筛选",
            "schema": {"type": "integer"}
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "搜索关键词（文件名、描述等）",
            "schema": {"type": "string"}
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "排序字段",
            "schema": {"type": "string", "enum": ["created_at", "updated_at", "file_size", "file_name"], "default": "created_at"}
          },
          {
            "name": "order",
            "in": "query",
            "description": "排序方向",
            "schema": {"type": "string", "enum": ["ASC", "DESC"], "default": "DESC"}
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {"$ref": "#/components/schemas/PaginationResponse"},
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "files": {
                              "type": "array",
                              "items": {"$ref": "#/components/schemas/File"}
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/files/{id}": {
      "get": {
        "tags": ["文件管理"],
        "summary": "获取文件详情",
        "description": "根据ID获取文件详细信息（需要管理员权限）",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "文件ID",
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {"$ref": "#/components/schemas/File"}
                  }
                }
              }
            }
          },
          "404": {
            "description": "文件不存在",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          }
        }
      },
      "put": {
        "tags": ["文件管理"],
        "summary": "更新文件信息",
        "description": "更新文件的元数据信息（需要管理员权限）",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "文件ID",
            "schema": {"type": "integer"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "description": "文件描述",
                    "example": "更新后的文件描述"
                  },
                  "tags": {
                    "type": "array",
                    "description": "文件标签",
                    "items": {"type": "string"},
                    "example": ["技术", "文档", "更新"]
                  },
                  "related_id": {
                    "type": "integer",
                    "description": "关联ID",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/SuccessResponse"}
              }
            }
          },
          "404": {
            "description": "文件不存在",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["文件管理"],
        "summary": "删除文件",
        "description": "删除指定文件（需要管理员权限）",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "文件ID",
            "schema": {"type": "integer"}
          },
          {
            "name": "hardDelete",
            "in": "query",
            "description": "是否物理删除文件",
            "schema": {"type": "boolean", "default": false}
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/SuccessResponse"}
              }
            }
          },
          "404": {
            "description": "文件不存在",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          }
        }
      }
    },
...
  },
...
} 